Implement LevelDB support in Bitcoin.
This accelerates the tx index database significantly. See the
discussion on bitcoin-dev for details.
