+++ src/bitcoinrpc.cpp
@@ -3,6 +3,11 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "main.h"
+#include "wallet.h"
+#include "txdb.h"
+#include "walletdb.h"
+#include "net.h"
 #include "init.h"
 #include "util.h"
 #include "sync.h"
