+++ src/main.h
@@ -592,7 +592,7 @@ public:
     {
         CAutoFile filein = CAutoFile(OpenBlockFile(pos.nFile, 0, pfileRet ? "rb+" : "rb"), SER_DISK, CLIENT_VERSION);
         if (!filein)
-            return error("CTransaction::ReadFromDisk() : OpenBlockFile failed");
+            return error("CTransaction::ReadFromDisk() : OpenBlockFile failed: %s", pos.ToString().c_str());
 
         // Read transaction
         if (fseek(filein, pos.nTxPos, SEEK_SET) != 0)
