+++ src/makefile.linux-mingw
@@ -6,6 +6,7 @@ DEPSDIR:=/usr/i586-mingw32msvc
 
 USE_UPNP:=0
 USE_IPV6:=1
+USE_LEVELDB:=1
 
 INCLUDEPATHS= \
  -I"$(CURDIR)" \
@@ -60,7 +61,6 @@ OBJS= \
     obj/addrman.o \
     obj/crypter.o \
     obj/key.o \
-    obj/db.o \
     obj/init.o \
     obj/irc.o \
     obj/keystore.o \
@@ -82,6 +82,20 @@ OBJS= \
     obj/noui.o
 
 all: bitcoind.exe
+
+#
+# LevelDB support
+#
+ifdef USE_LEVELDB
+LIBS += $(CURDIR)/leveldb-1.5.0/libleveldb.lib
+DEFS += -I"$(CURDIR)/leveldb-1.5.0/include" -DUSE_LEVELDB
+OBJS += obj/txdb-leveldb.o
+leveldb-1.5.0/libleveldb.lib:
+	@echo "Building LevelDB ..."; cd leveldb-1.5.0; TARGET_OS=BitcoinCrossCompileMingW make; cd ..
+obj/txdb-leveldb.o: leveldb-1.5.0/libleveldb.a
+else
+OBJS += obj/txdb-bdb.o
+endif
 
 obj/build.h: FORCE
 	/bin/sh ../share/genbuild.sh obj/build.h
@@ -109,5 +123,6 @@ clean:
 	-rm -f obj-test/*.o
 	-rm -f test_bitcoin.exe
 	-rm -f src/build.h
+	-cd leveldb-1.5.0; make clean; cd ..
 
 FORCE:
