+++ src/rpcmining.cpp
@@ -5,6 +5,7 @@
 
 #include "main.h"
 #include "db.h"
+#include "txdb.h"
 #include "init.h"
 #include "bitcoinrpc.h"
